requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-identity>=1.12.0
azure-functions
databricks-sql-connector
# functions_framework >> used for google cloud functions